// rms_merged_v2/src/types.ts
export type CommodityId =
  | 'wheat'
  | 'corn'
  | 'soybean'
  | 'soybean-oil'
  | 'palm-oil'
  | 'sugar'
  | 'potato-starch'
  | 'tapioca-starch';